src/requirements.txt
proxmoxer
requests
discord.py
urllib3
pytest
pytest-cov
pytest-mock
pytest-asyncio